+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.21.5] - 2023-05-03
+- When the "Strip Unity Version" build flag is used, the Unity Version is stripped for the archive file header and serialized files in the bundle.
+- Fixed issue when changing dependencies of an asset change from being explicit to implicit and non-recursive dependencies enabled, and cache is available. Can result in incorrect depdencies being built incorrectly.
+- Improvements to the documentation
+- Fixed bug where source textures were getting included in the build when a SpriteAtlas was also included
+
 ## [1.21.3] - 2023-01-09
 - Changed calculating AssetDependency to generate dependencies by Object instead of per Asset